Began implementing Jira interface.
