+++ report/html_report.go
@@ -1,179 +1,30 @@
 package report
 
 import (
-	"encoding/json"
 	"fmt"
 	"html/template"
 	"os"
-	"regexp"
 	"runtime"
 	"time"
 
 	"github.com/cbush06/kosher/config"
-	"github.com/cbush06/kosher/fs"
+
+	"github.com/cbush06/kosher/common"
 	"github.com/cbush06/kosher/resources/reporttemplates"
 	"github.com/spf13/afero"
 )
 
 const errMsg = "Error encountered while generating HTML report: %s"
 
-var leadingWhitespace = regexp.MustCompile(`(?m)^(?:\s+)(.*)`)
-
-// cukeComment is any single-line comment.
-type cukeComment struct {
-	Value string `json:"value"`
-	Line  int    `json:"line"`
-}
-
-// cukeDocstring is a docstring (multi-line comment).
-type cukeDocstring struct {
-	Value       string `json:"value"`
-	ContentType string `json:"content_type"`
-	Line        int    `json:"line"`
-}
-
-// cuckeTag is any tag added to a feature or scenario.
-type cukeTag struct {
-	Name string `json:"name"`
-	Line int    `json:"line"`
-}
-
-// cukeResult is the result of executing a step.
-type cukeResult struct {
-	Status   string `json:"status"`
-	Error    string `json:"error_message,omitempty"`
-	Duration *int   `json:"duration,omitempty"`
-}
-
-func (r *cukeResult) GetDurationInSeconds() string {
-	if r.Duration != nil {
-		return fmt.Sprintf("%0.2fs", time.Duration(*r.Duration).Seconds())
-	}
-	return ""
-}
-
-// cuckeMatch is the method matched to a step.
-type cukeMatch struct {
-	Location string `json:"location"`
-}
-
-// cukeEmbedding is any file or binary content attached to a step.
-type cukeEmbedding struct {
-	MimeType string `json:"mime_type"`
-	Data     string `json:"data"`
-}
-
-// cukeStep is a single step in a cukeElement.
-type cukeStep struct {
-	Keyword    string              `json:"keyword"`
-	Name       string              `json:"name"`
-	Line       int                 `json:"line"`
-	Docstring  *cukeDocstring      `json:"doc_string,omitempty"`
-	Match      cukeMatch           `json:"match"`
-	Result     cukeResult          `json:"result"`
-	Embeddings []cukeEmbedding     `json:"embeddings,omitempty"`
-	DataTable  []*cukeDataTableRow `json:"rows,omitempty"`
-}
-
-// cukeDataTableRow represents a row in a DataTable owned by a step
-type cukeDataTableRow struct {
-	Cells []string `json:"cells"`
-}
-
-// cukeElement represents any block nested within a Feature:
-//		* Background
-//		* Scenario
-//		* Scneario Outline
-type cukeElement struct {
-	ID           string     `json:"id"`
-	Keyword      string     `json:"keyword"`
-	Name         string     `json:"name"`
-	Description  string     `json:"description"`
-	Line         int        `json:"line"`
-	Type         string     `json:"type"`
-	Tags         []cukeTag  `json:"tags,omitempty"`
-	Steps        []cukeStep `json:"steps,omitempty"`
-	StepsPassed  int        `json:"-"`
-	StepsFailed  int        `json:"-"`
-	StepsPending int        `json:"-"`
-	StepsSkipped int        `json:"-"`
-}
-
-func (e *cukeElement) GetTrimmedDescription() string {
-	return leadingWhitespace.ReplaceAllString(e.Description, "$1")
-}
-
-// cukeFeature is a single feature in JSONReport.
-type cukeFeature struct {
-	URI             string        `json:"uri"`
-	ID              string        `json:"id"`
-	Keyword         string        `json:"keyword"`
-	Name            string        `json:"name"`
-	Description     string        `json:"description"`
-	Line            int           `json:"line"`
-	Comments        []cukeComment `json:"comments,omitempty"`
-	Tags            []cukeTag     `json:"tags,omitempty"`
-	Elements        []cukeElement `json:"elements,omitempty"`
-	ElementsPassed  int           `json:"-"`
-	ElementsFailed  int           `json:"-"`
-	ElementsPending int           `json:"-"`
-	StepsPassed     int           `json:"-"`
-	StepsFailed     int           `json:"-"`
-	StepsPending    int           `json:"-"`
-	StepsSkipped    int           `json:"-"`
-}
-
-// GetTrimmedDescription returns the features description after removing leading whitespace from each line.
-func (f *cukeFeature) GetTrimmedDescription() string {
-	return leadingWhitespace.ReplaceAllString(f.Description, "$1")
-}
-
 // HTMLReport holds the jsonResults of a test execution
 type HTMLReport struct {
-	fileSystem      *fs.Fs
-	settings        *config.Settings
-	jsonResults     []byte
-	Features        []cukeFeature
-	ProjectName     string
-	AppVersion      string
-	Environment     string
-	Browser         string
-	Platform        string
-	RunTime         time.Duration
-	OS              string
-	Timestamp       string
-	ElementsPassed  int
-	ElementsFailed  int
-	ElementsPending int
-	StepsPassed     int
-	StepsFailed     int
-	StepsPending    int
-	StepsSkipped    int
-	TotalElements   int
-	TotalSteps      int
+	CucumberReport
+	jsonResults []byte
 }
 
-func newHTMLReport(s *config.Settings, f *fs.Fs) *HTMLReport {
+func newHTMLReport(s *config.Settings) *HTMLReport {
 	return &HTMLReport{
-		fileSystem:      f,
-		settings:        s,
-		ProjectName:     s.Settings.GetString("projectName"),
-		AppVersion:      s.Settings.GetString("appVersion"),
-		Environment:     s.Settings.GetString("environment"),
-		Browser:         s.Settings.GetString("driver"),
-		Platform:        s.Settings.GetString("platform"),
-		RunTime:         time.Duration(0),
-		OS:              "",
-		Timestamp:       "",
-		ElementsPassed:  0,
-		ElementsFailed:  0,
-		ElementsPending: 0,
-		StepsPassed:     0,
-		StepsFailed:     0,
-		StepsPending:    0,
-		StepsSkipped:    0,
-		TotalElements:   0,
-		TotalSteps:      0,
+		CucumberReport: NewCucumberReport(s),
 	}
 }
 
@@ -195,8 +46,8 @@ func (r *HTMLReport) Process() error {
 	)
 
 	const (
-		resultsHTMLFile = "results.html"
-		resultsJSONFile = "results.json"
+		resultsHTMLFile = common.ResultsHTMLFile
+		resultsJSONFile = common.ResultsJSONFile
 	)
 
 	reportFormat := r.settings.Settings.GetString("reportFormat")
@@ -213,13 +64,13 @@ func (r *HTMLReport) Process() error {
 		return fmt.Errorf(errMsg, "attempt made to generate HTML report with unrecognized template")
 	}
 
-	if err = r.unmarshallJSON(); err != nil {
+	if err = r.UnmarshallJSON(r.jsonResults); err != nil {
 		return err
 	}
 
 	// write HTML report file
-	filePath, _ := r.fileSystem.ResultsDir.RealPath(resultsHTMLFile)
-	if fileHandle, err = r.fileSystem.ResultsDir.OpenFile(resultsHTMLFile, os.O_CREATE|os.O_WRONLY|os.O_TRUNC, 0664); err != nil {
+	filePath, _ := r.settings.FileSystem.ResultsDir.RealPath(resultsHTMLFile)
+	if fileHandle, err = r.settings.FileSystem.ResultsDir.OpenFile(resultsHTMLFile, os.O_CREATE|os.O_WRONLY|os.O_TRUNC, 0664); err != nil {
 		return fmt.Errorf(fmt.Sprintf(errMsg, "failed to open results file [%s]: %s"), filePath, err)
 	}
 	if err = templ.Execute(fileHandle, r); err != nil {
@@ -230,8 +81,8 @@ func (r *HTMLReport) Process() error {
 	}
 
 	// write JSON results file
-	filePath, _ = r.fileSystem.ResultsDir.RealPath(resultsJSONFile)
-	if fileHandle, err = r.fileSystem.ResultsDir.OpenFile(resultsJSONFile, os.O_CREATE|os.O_WRONLY|os.O_TRUNC, 0664); err != nil {
+	filePath, _ = r.settings.FileSystem.ResultsDir.RealPath(resultsJSONFile)
+	if fileHandle, err = r.settings.FileSystem.ResultsDir.OpenFile(resultsJSONFile, os.O_CREATE|os.O_WRONLY|os.O_TRUNC, 0664); err != nil {
 		return fmt.Errorf(fmt.Sprintf(errMsg, "failed to open results file [%s]: %s"), filePath, err)
 	}
 	if writeLen, err = fileHandle.Write(r.jsonResults); err != nil {
@@ -241,71 +92,6 @@ func (r *HTMLReport) Process() error {
 	}
 	if err = fileHandle.Close(); err != nil {
 		return fmt.Errorf(fmt.Sprintf(errMsg, "failed to close report file [%s]: %s"), filePath, err)
-	}
-
-	return nil
-}
-
-func (r *HTMLReport) unmarshallJSON() error {
-	if err := json.Unmarshal(r.jsonResults, &r.Features); err != nil {
-		return fmt.Errorf(fmt.Sprintf(errMsg, "failed to parse JSON results of test execution: %s"), err)
-	}
-
-	for f := 0; f < len(r.Features); f++ {
-		feature := &r.Features[f]
-
-		for e := 0; e < len(feature.Elements); e++ {
-			element := &feature.Elements[e]
-
-			for s := 0; s < len(element.Steps); s++ {
-				step := &element.Steps[s]
-
-				if step.DataTable == nil {
-					step.DataTable = make([]*cukeDataTableRow, 0)
-				}
-
-				switch step.Result.Status {
-				case "passed":
-					element.StepsPassed++
-				case "failed":
-					element.StepsFailed++
-				case "skipped":
-					element.StepsSkipped++
-				case "undefined":
-					element.StepsPending++
-				}
-
-				if step.Result.Duration != nil {
-					r.RunTime += time.Duration(*step.Result.Duration)
-				}
-
-				r.TotalSteps++
-			}
-
-			feature.StepsPassed += element.StepsPassed
-			feature.StepsFailed += element.StepsFailed
-			feature.StepsSkipped += element.StepsSkipped
-			feature.StepsPending += element.StepsPending
-
-			r.StepsPassed += element.StepsPassed
-			r.StepsFailed += element.StepsFailed
-			r.StepsSkipped += element.StepsSkipped
-			r.StepsPending += element.StepsPending
-
-			if element.StepsFailed > 0 {
-				feature.ElementsFailed++
-			} else if element.StepsPending > 0 {
-				feature.ElementsPending++
-			} else {
-				feature.ElementsPassed++
-			}
-
-			r.TotalElements++
-		}
-
-		r.ElementsPassed += feature.ElementsPassed
-		r.ElementsFailed += feature.ElementsFailed
-		r.ElementsPending += feature.ElementsPending
 	}
 
 	return nil
